+++ app/src/main/java/com/adammcneilly/reader/MainActivity.kt
@@ -1,4 +1,4 @@
-package template
+package com.adammcneilly.reader
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,7 +8,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import template.theme.TemplateTheme
+import com.adammcneilly.reader.theme.ReaderTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -17,7 +17,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             enableEdgeToEdge()
 
-            TemplateTheme {
+            ReaderTheme {
                 Surface(
                     color = MaterialTheme.colorScheme.background,
                 ) {
