+++ app/src/main/java/com/adammcneilly/reader/ReaderApp.kt
@@ -1,7 +1,7 @@
-package template
+package com.adammcneilly.reader
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class TemplateApp : Application()
+class ReaderApp : Application()
