+++ app/src/main/java/com/adammcneilly/reader/theme/Shape.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.adammcneilly.reader.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
