+++ gradle/libs.versions.toml
@@ -24,8 +24,6 @@ paparazzi = "1.3.5"
 retrofit = "3.0.0"
 room = "2.7.1"
 sortDependencies = "0.14"
-xr = "1.0.0-alpha04"
-xr-material = "1.0.0-alpha08"
 
 [libraries]
 android-material = { module = "com.google.android.material:material", version.ref = "material" }
@@ -37,9 +35,6 @@ androidx-room-compiler = { module = "androidx.room:room-compiler", version.ref =
 androidx-room-runtime = { module = "androidx.room:room-runtime", version.ref = "room" }
 androidx-test-espresso-core = { module = "androidx.test.espresso:espresso-core", version.ref = "espresso" }
 androidx-test-junit = { module = "androidx.test.ext:junit", version.ref = "androidxTest" }
-androidx-xr-compose = { module = "androidx.xr.compose:compose", version.ref = "xr" }
-androidx-xr-compose-material3 = { module = "androidx.xr.compose.material3:material3", version.ref = "xr-material" }
-androidx-xr-scenecore = { module = "androidx.xr.scenecore:scenecore", version.ref = "xr" }
 coil-compose = { module = "io.coil-kt.coil3:coil-compose", version.ref = "coil" }
 coil-okhttp = { module = "io.coil-kt.coil3:coil-network-okhttp", version.ref = "coil" }
 compose-bom = { group = "androidx.compose", name = "compose-bom", version.ref = "composeBom" }
@@ -77,4 +72,3 @@ kotlinter = { id = "org.jmailen.kotlinter", version.ref = "kotlinter" }
 square-sort-dependencies = { id = "com.squareup.sort-dependencies", version.ref = "sortDependencies" }
 
 [bundles]
-androidx-xr = ["androidx-xr-compose", "androidx-xr-compose-material3", "androidx-xr-scenecore"]
